+++ test/unit/v1/reactions/test_parse_arrhenius.cpp
@@ -9,7 +9,7 @@ TEST(ParserBase, CanParseValidArrheniusReaction)
   std::vector<std::string> extensions = { ".json", ".yaml" };
   for (auto& extension : extensions)
   {
-    auto parsed = parse(std::string("v1_unit_configs/reactions/arrhenius/valid") + extension);
+    auto parsed = Parse(std::string("v1_unit_configs/reactions/arrhenius/valid") + extension);
     EXPECT_TRUE(parsed);
     Mechanism mechanism = *parsed;
 
@@ -73,7 +73,7 @@ TEST(ParserBase, ArrheniusDetectsUnknownSpecies)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/arrhenius/unknown_species") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 2);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::ReactionRequiresUnknownSpecies);
@@ -90,7 +90,7 @@ TEST(ParserBase, ArrheniusDetectsMutuallyExclusiveOptions)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/arrhenius/mutually_exclusive") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::MutuallyExclusiveOption);
@@ -107,7 +107,7 @@ TEST(ParserBase, ArrheniusDetectsBadReactionComponent)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/arrhenius/bad_reaction_component") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 2);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::RequiredKeyNotFound);
@@ -125,7 +125,7 @@ TEST(ParserBase, ArrheniusDetectsUnknownPhase)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/arrhenius/missing_phase") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::UnknownPhase);