+++ test/unit/v1/reactions/test_parse_first_order_loss.cpp
@@ -9,7 +9,7 @@ TEST(ParserBase, CanParseValidFirstOrderLossReaction)
   std::vector<std::string> extensions = { ".json", ".yaml" };
   for (auto& extension : extensions)
   {
-    auto parsed = parse(std::string("v1_unit_configs/reactions/first_order_loss/valid") + extension);
+    auto parsed = Parse(std::string("v1_unit_configs/reactions/first_order_loss/valid") + extension);
     EXPECT_TRUE(parsed);
     Mechanism mechanism = *parsed;
 
@@ -37,7 +37,7 @@ TEST(ParserBase, FirstOrderLossDetectsUnknownSpecies)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/first_order_loss/unknown_species") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::ReactionRequiresUnknownSpecies);
@@ -54,7 +54,7 @@ TEST(ParserBase, FirstOrderLossDetectsBadReactionComponent)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/first_order_loss/bad_reaction_component") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 2);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::RequiredKeyNotFound);
@@ -72,7 +72,7 @@ TEST(ParserBase, FirstOrderLossDetectsUnknownPhase)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/first_order_loss/missing_phase") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::UnknownPhase);
@@ -89,7 +89,7 @@ TEST(ParserBase, FirstOrderLossDetectsMoreThanOneSpecies)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/first_order_loss/too_many_reactants") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::TooManyReactionComponents);
@@ -105,7 +105,7 @@ TEST(ParserBase, CanParseValidFirstOrderLossReactionWithProducts)
   std::vector<std::string> extensions = { ".json", ".yaml" };
   for (auto& extension : extensions)
   {
-    auto parsed = parse(std::string("v1_unit_configs/reactions/first_order_loss/products") + extension);
+    auto parsed = Parse(std::string("v1_unit_configs/reactions/first_order_loss/products") + extension);
     EXPECT_TRUE(parsed);
     Mechanism mechanism = *parsed;