+++ test/unit/v1/reactions/test_parse_surface.cpp
@@ -9,7 +9,7 @@ TEST(ParserBase, CanParseValidSurfaceReaction)
   std::vector<std::string> extensions = { ".json", ".yaml" };
   for (auto& extension : extensions)
   {
-    auto parsed = parse(std::string("v1_unit_configs/reactions/surface/valid") + extension);
+    auto parsed = Parse(std::string("v1_unit_configs/reactions/surface/valid") + extension);
     EXPECT_TRUE(parsed);
     Mechanism mechanism = *parsed;
 
@@ -48,7 +48,7 @@ TEST(ParserBase, SurfaceDetectsUnknownSpecies)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/surface/unknown_species") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::ReactionRequiresUnknownSpecies);
@@ -65,7 +65,7 @@ TEST(ParserBase, SurfaceDetectsBadReactionComponent)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/surface/bad_reaction_component") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 2);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::RequiredKeyNotFound);
@@ -83,7 +83,7 @@ TEST(ParserBase, SurfaceDetectsUnknownGasPhase)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/surface/missing_gas_phase") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::UnknownPhase);