+++ test/unit/v1/reactions/test_parse_taylor_series.cpp
@@ -9,7 +9,7 @@ TEST(ParserBase, CanParseValidTaylorSeriesReaction)
   std::vector<std::string> extensions = { ".json", ".yaml" };
   for (auto& extension : extensions)
   {
-    auto parsed = parse(std::string("v1_unit_configs/reactions/taylor_series/valid") + extension);
+    auto parsed = Parse(std::string("v1_unit_configs/reactions/taylor_series/valid") + extension);
     EXPECT_TRUE(parsed);
     Mechanism mechanism = *parsed;
 
@@ -81,7 +81,7 @@ TEST(ParserBase, TaylorSeriesDetectsUnknownSpecies)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/taylor_series/unknown_species") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::ReactionRequiresUnknownSpecies);
@@ -98,7 +98,7 @@ TEST(ParserBase, TaylorSeriesDetectsMutuallyExclusiveOptions)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/taylor_series/mutually_exclusive") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::MutuallyExclusiveOption);
@@ -115,7 +115,7 @@ TEST(ParserBase, TaylorSeriesDetectsBadReactionComponent)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/taylor_series/bad_reaction_component") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 2);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::RequiredKeyNotFound);
@@ -133,7 +133,7 @@ TEST(ParserBase, TaylorSeriesDetectsUnknownPhase)
   for (auto& extension : extensions)
   {
     std::string file = std::string("v1_unit_configs/reactions/taylor_series/missing_phase") + extension;
-    auto parsed = parse(file);
+    auto parsed = Parse(file);
     EXPECT_FALSE(parsed);
     EXPECT_EQ(parsed.error().size(), 1);
     EXPECT_EQ(parsed.error()[0].first, ErrorCode::UnknownPhase);